fix(Whole Project): Renamed files and folders.
Renamed all files and folders according to Dart preferences.

fixes #160
