+++ lib/Providers/theme_provider.dart
@@ -1,4 +1,4 @@
-import 'package:doclense/Services/shared_preferences.dart';
+import 'package:doclense/services/shared_preferences.dart';
 import 'package:flutter/material.dart';
 
 class DarkThemeProvider with ChangeNotifier {
