+++ lib/Services/search_service.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 import 'package:open_file/open_file.dart';
-// import 'package:doclense/Constants/theme_constants.dart';
+// import 'package:doclense/constants/theme_constants.dart';
 
 class SearchService extends SearchDelegate<String> {
   List<dynamic> files = Hive.box('pdfs').getAt(0) as List<dynamic>;
@@ -79,7 +79,7 @@ class SearchService extends SearchDelegate<String> {
       itemBuilder: (context, index) => ListTile(
         onTap: () {
           final List<dynamic> finalList = [];
-          for(int i=0; i<files.length; i++){
+          for (int i = 0; i < files.length; i++) {
             finalList.add(files[i][0]);
           }
           final String fileLocation =
@@ -104,9 +104,7 @@ class SearchService extends SearchDelegate<String> {
   @override
   Widget buildSuggestions(BuildContext context) {
     for (int i = 0; i < files.length; i++) {
-      pdfNames.add(Hive.box('pdfs').getAt(0)[i][0]
-          .split('/')
-          .last);
+      pdfNames.add(Hive.box('pdfs').getAt(0)[i][0].split('/').last);
     }
 
     final int remove = (3 * pdfNames.length / 16).floor();
@@ -118,44 +116,35 @@ class SearchService extends SearchDelegate<String> {
     final suggestedFiles = query.isEmpty
         ? recentFiles
         : pdfNames
-        .where((p) => p.startsWith(query.toLowerCase()) as bool)
-        .toList();
+            .where((p) => p.startsWith(query.toLowerCase()) as bool)
+            .toList();
 
     // print(files.length);
     print(pdfNames);
     // print(suggestedFiles);
 
     return ListView.builder(
-      itemBuilder: (context, index) =>
-          ListTile(
-            onTap: () {
-              final List<dynamic> finalList = [];
-              for (int i = 0; i < files.length; i++) {
-                finalList.add(files[i][0]);
-              }
-              final String fileLocation =
-                  '${finalList.where((element) =>
-              element
-                  .toString()
-                  .split('/')
-                  .last == suggestedFiles[index])}';
-              print(finalList);
+      itemBuilder: (context, index) => ListTile(
+        onTap: () {
+          final List<dynamic> finalList = [];
+          for (int i = 0; i < files.length; i++) {
+            finalList.add(files[i][0]);
+          }
+          final String fileLocation =
+              '${finalList.where((element) => element.toString().split('/').last == suggestedFiles[index])}';
+          print(finalList);
 
-              final String fileLoc = fileLocation
-                  .split('(')
-                  .last
-                  .split(')')
-                  .first;
+          final String fileLoc = fileLocation.split('(').last.split(')').first;
 
-              // query = fileLoc.split('/').last;
+          // query = fileLoc.split('/').last;
 
-              OpenFile.open(fileLoc);
-            },
-            leading: Icon(query.isEmpty
-                ? Icons.youtube_searched_for_rounded
-                : Icons.picture_as_pdf_rounded),
-            title: Text(suggestedFiles[index] as String),
-          ),
+          OpenFile.open(fileLoc);
+        },
+        leading: Icon(query.isEmpty
+            ? Icons.youtube_searched_for_rounded
+            : Icons.picture_as_pdf_rounded),
+        title: Text(suggestedFiles[index] as String),
+      ),
       itemCount: suggestedFiles.length,
     );
   }
