+++ lib/home.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:io';
-import 'package:doclense/Services/search_service.dart';
+import 'package:doclense/services/search_service.dart';
 import 'package:doclense/starred_documents.dart';
 
 // import 'package:path/path.dart' as path;
@@ -21,8 +21,8 @@ import 'package:quick_actions/quick_actions.dart';
 import 'package:share/share.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import 'Providers/image_list.dart';
-import 'Providers/theme_provider.dart';
+import 'providers/image_list.dart';
+import 'providers/theme_provider.dart';
 import 'about.dart';
 import 'image_view.dart';
 import 'main_drawer.dart';
@@ -560,7 +560,8 @@ class _HomeState extends State<Home> {
                                           : Colors.grey,
                                     ),
                                     onPressed: () async {
-                                      final status = await Permission.storage.status;
+                                      final status =
+                                          await Permission.storage.status;
                                       if (!status.isGranted) {
                                         await Permission.storage.request();
                                       }
