+++ lib/settings/settings.dart
@@ -1,11 +1,11 @@
-// import 'package:doclense/Constants/theme_constants.dart';
+// import 'package:doclense/constants/theme_constants.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:package_info/package_info.dart';
 import 'package:provider/provider.dart';
 import 'package:wiredash/wiredash.dart';
 
-import '../Providers/theme_provider.dart';
+import '../providers/theme_provider.dart';
 import '../main_drawer.dart';
 import './contact_developers.dart';
 import 'setting_text.dart';
