+++ lib/starred_documents.dart
@@ -8,7 +8,7 @@ import 'package:open_file/open_file.dart';
 import 'package:provider/provider.dart';
 import 'package:share/share.dart';
 
-import 'Providers/theme_provider.dart';
+import 'providers/theme_provider.dart';
 import 'main_drawer.dart';
 
 class Starred extends StatefulWidget {
@@ -40,17 +40,13 @@ class _StarredState extends State<Starred> {
       body: WatchBoxBuilder(
         box: Hive.box('starred'),
         builder: (context, starredBox) {
-          if (starredBox
-              .getAt(0)
-              .length == 0) {
+          if (starredBox.getAt(0).length == 0) {
             return const Center(
               child: Text("No PDFs Starred Yet !! "),
             );
           }
           return ListView.builder(
-            itemCount: starredBox
-                .getAt(0)
-                .length as int,
+            itemCount: starredBox.getAt(0).length as int,
             itemBuilder: (context, index) {
               return GestureDetector(
                 onTap: () {
@@ -61,8 +57,8 @@ class _StarredState extends State<Starred> {
                   padding: const EdgeInsets.all(15.0),
                   child: Card(
                     elevation: 5,
-                    color: themeChange.darkTheme ? Colors.grey[700] : Colors
-                        .white,
+                    color:
+                        themeChange.darkTheme ? Colors.grey[700] : Colors.white,
                     child: Row(
                       children: [
                         Column(
@@ -72,10 +68,7 @@ class _StarredState extends State<Starred> {
                               child: Image(
                                 image: FileImage(
                                     starredBox.getAt(0)[index][2] as File),
-                                width: MediaQuery
-                                    .of(context)
-                                    .size
-                                    .width / 6,
+                                width: MediaQuery.of(context).size.width / 6,
                               ),
                             )
                           ],
@@ -98,19 +91,14 @@ class _StarredState extends State<Starred> {
                             ),
                             SizedBox(
                                 height:
-                                MediaQuery
-                                    .of(context)
-                                    .size
-                                    .height * 0.01),
+                                    MediaQuery.of(context).size.height * 0.01),
                             Row(
                               children: [
                                 SizedBox(
-                                  width: (MediaQuery
-                                      .of(context)
-                                      .size
-                                      .width * MediaQuery
-                                      .of(context)
-                                      .devicePixelRatio) * 0.1,
+                                  width: (MediaQuery.of(context).size.width *
+                                          MediaQuery.of(context)
+                                              .devicePixelRatio) *
+                                      0.1,
                                 ),
                                 IconButton(
                                     icon: const Icon(Icons.share),
